refactor: Group features in folders
