+++ web/src/components/feature/results/MercuryWarning.tsx
