+++ web/src/components/feature/results/PhaseTable.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react'
 import { PhaseTable } from './PhaseTable'
 import '@testing-library/jest-dom/extend-expect'
-import { mockMultiflashResponse } from '../../setupTests'
+import { mockMultiflashResponse } from '../../../setupTests'
 
 test('renders without crashing and displaying correct values in table', async () => {
   render(
