+++ web/src/pages/Main.tsx
@@ -2,13 +2,13 @@ import styled from 'styled-components'
 import { Divider, Typography } from '@equinor/eds-core-react'
 import { Header } from '../components/common/Header'
 import { useEffect, useState } from 'react'
-import { MoleTable } from '../components/feature/MoleTable'
-import { CalculationInput } from '../components/feature/CalculationInput'
+import { MoleTable } from '../components/feature/results/MoleTable'
+import { CalculationInput } from '../components/feature/calculation_input/CalculationInput'
 import MercuryAPI from '../api/MercuryAPI'
 import { AxiosResponse } from 'axios'
 import { ComponentResponse, MultiflashResponse } from '../api/generated'
-import { PhaseTable } from '../components/feature/PhaseTable'
-import { MercuryWarning } from '../components/feature/MercuryWarning'
+import { PhaseTable } from '../components/feature/results/PhaseTable'
+import { MercuryWarning } from '../components/feature/results/MercuryWarning'
 import { TComponentProperties, TComponentRatios, TFeedFlow } from '../types'
 
 const Results = styled.div`
